refactored helper class to standard files
