+++ Controller/Admin/ModuleList.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace OxCom\ModuleInternals\Controller\Admin;
+
+/**
+ * Class Metadata
+ * extendeing OxidEsales\Eshop\Core\Module\ModuleList
+ */
+class ModuleList extends ModuleList_parent
+{
+
+}
