+++ Core/Module.php
@@ -1,0 +1,99 @@
+<?php
+namespace OxCom\ModuleInternals\Core;
+
+use \OxidEsales\Eshop\Core\DatabaseProvider as DatabaseProvider;
+use \OxidEsales\Eshop\Core\Registry as Registry;
+use \OxidEsales\Eshop\Core\Module\ModuleList as ModuleList;
+
+/**
+ * Class Metadata
+ * extendeing OxidEsales\Eshop\Core\Module\Module as Module
+ */
+class Module extends Module_parent
+{
+    /**
+     * Get template blocks defined in database.
+     *
+     * @return array
+     */
+    public function getModuleBlocks()
+    {
+        $aResults = DatabaseProvider::getDb(DatabaseProvider::FETCH_MODE_ASSOC)->select(
+            'SELECT * FROM oxtplblocks WHERE oxModule = ? AND oxshopid = ?',
+            [$this->getId(), Registry::getConfig()->getShopId()]
+        );
+
+        return $aResults->fetchAll();
+    }
+
+    /**
+     * Get module settings defined in database.
+     *
+     * @return array
+     */
+    public function getModuleSettings()
+    {
+        $aResult = DatabaseProvider::getDb(DatabaseProvider::FETCH_MODE_ASSOC)->select(
+            'SELECT * FROM oxconfig WHERE oxModule = ? AND oxshopid = ?',
+            [sprintf('module:%s', $this->getId()), Registry::getConfig()->getShopId()]
+        );
+
+        return $aResult->fetchAll();
+    }
+
+    /**
+     * Check supported metadata version
+     *
+     * @return bool
+     */
+    public function isMetadataSupported()
+    {
+        $sMetadataVersion = $this->getMetaDataVersion();
+
+        $sLatestVersion = '1.0';
+
+        if (method_exists('oxModuleList', 'getModuleVersions') || method_exists('oxModule', 'getModuleEvents')) {
+            $sLatestVersion = '1.1';
+        }
+
+        if (method_exists(ModuleList::class, 'getModuleConfigParametersByKey')) {
+            $sLatestVersion = '2.0';
+        }
+
+        return version_compare($sLatestVersion, $sMetadataVersion) >= 0;
+    }
+
+    /**
+     * check if current metadata version is $sVersion
+     *
+     * @param $sVersion
+     * @return bool
+     */
+    public function checkMetadataVersion($sVersion){
+        return version_compare($this->getMetaDataVersion(), $sVersion) == 0;
+    }
+
+    /**
+     * Returns array of module files / entries in metadata.php
+     * like 'template' => / 'blocks' => ....
+     * possible entries
+     *
+     * ModuleList::MODULE_KEY_FILES = 'files'
+     * ModuleList::MODULE_KEY_CONTROLLERS = 'controllers'
+     * ModuleList::MODULE_KEY_EVENTS => 'events'
+     * ModuleList::MODULE_KEY_VERSIONS => module version
+     *
+     * @return array
+     */
+    public function getModuleEntries($sType)
+    {
+        $aReturn = [];
+        $aList = oxNew(ModuleList::class)->getModuleConfigParametersByKey($sType);
+
+        if (isset($aList[$this->getId()])) {
+            $aReturn = $aList[$this->getId()];
+        }
+
+        return $aReturn;
+    }
+}
